NMake Makefiles: Split outdir by toolset/pdb version
This reduces the likelihood of accidently mixing binaries linked against
different CRT versions.

Visual Studio 2015 through 2019 link to the same CRT/C++ runtime DLL, so
they are interchangeable as long as the system has the CRT/C++RT
DLLs installed that added support for the later toolsets, so use the PDB
version instead.