NMake Makefiles: Stop using gendef.exe by default
Define CAIROMM_API as __declspec(dllexport) (when building) or
__declspec(dllimport) (when using) by default, unless we decide that we
want to continue using gendef.exe for compatibility reasons, by passing
in BUILD_COMPAT_LIB in the NMake command line.  This means that gendef
is not used unless we explicitly request to do so.

We may be using the cairomm binaries on older programs that we aren't
able to rebuild, so allow gendef to be still used if it is desired,
although using __declspec(dllexport) and __declspec(dllimport) will be
the default.