NMake Makefiles: Distinguish between MSVC 2015, 2017 and 2019
It was found that we cannot completely rely on the fact that Visual
Studio 2015~2019 tried very hard to be binary compatible, as there
could be corner cases when linking against cairomm built with Visual
Studio 2015 with builds done by Visual Studio 2017 and 2019 where
the code could fail to link and the DLLs are therefore not
ABI-compatible.  Note that the libsigc++ DLLs, however, are ABI
compatible between these 3 Visual Studio versions.

As a result, for the DLL and LIB names, use 'vc140' for Visual Studio
2015 builds, 'vc141' for Visual Studio 2017 builds and 'vc142' for
Visual Studio 2019 builds, according to the toolset versions as defined
by Microsoft.

For people that may have previously built cairomm with Visual
Studio 2017 or 2019, which had 'vc140' in the built .lib and DLL, an NMake
option 'USE_COMPAT_LIBS' is added to make building such binaries with
'vc140' easier, if needed.