MSVC_NMake/README: Add note on mixing between MSVC 2015~2019
Let people know that it is recommended to use the same Visual Studio
version, even when using 2015, 2017 or 2019, to build software that
depends on cairomm.

Also note a new NMake option that allows creating MSVC 2015-style DLL
and .lib filenames if that is needed, but note that is generally not
recommended.