NMake Makefiles: Fix builds from release tarballs
This avoids needlessly enerating cairommconfig.h and cairomm.rc, and when we do
need to generate these files, we ensure that they always have the right version
info in them.

Also streamline the build process that this file generating is done in the
'all' target, so that one does not need to use the -prep-git-build' target
beforehand.