cairomm/exception.h: Export Cairo::logic_error selectively
Since this class derives std::logic_error, don't export the class as a whole
but instead export its member methods as necessary on Windows.  This will make
the compiled DLL to not be locked in a single compiler/STL version, and will
remove the compiler warnings C4251, C4273 and C4275 on Visual Studio.