meson: Reorganize warnings on Visual Studio builds
We can now drop the ignores for warnings C4251, C4273 and C4275 along with
C4530, since our code and compiler flags should now make us free of them,
since we use /EHsc to build our code and we are clean of classes that we
export as a whole that derives from std::xxx classes.

Ignore warning C4800, and warning C4127 in the examples, since these warnings
are really spurious and safe to ignore as a whole.